refactor(docs): narrow OG-card exports and collapse render indirection
- Demote `BORDER`, `META_VALUE`, `PANEL_FILL`, `metaRow`, and `BG` in `parts.ts` plus `buildLandingCard` in `landing.ts` to file-private
- Extract the byte-identical outer-card wrapper into a shared `cardShell(...children)` helper in `parts.ts`
- Collapse `render.ts` into `template.ts`, moving `renderPage` next to a now-private `buildCard` for symmetry with `landing.ts`
- Strip the unused `slug` field from `OgPage` and its four constructor returns
- Replace `titleBlock`'s imperative children array with a conditional spread on `createElement`'s rest args
- Inline the single-use `left` binding in `titleArtwork` and adopt `src` / `fonts` object shorthand
- Drop the dead `gap = 24` default on `metaRow` since both callers pass it
